Extract validation functions into carina-core::validation module
Move validate_resources, validate_resource_ref_types, is_string_compatible_type,
validate_provider_config, validate_module_calls, and validate_module_arg_type
from carina-cli into carina-core::validation. CLI thin wrappers delegate to core
functions, passing schemas and factories as parameters.

Part of #328.
