+++ carina-core/src/lib.rs
@@ -16,3 +16,4 @@ pub mod resolver;
 pub mod resource;
 pub mod schema;
 pub mod utils;
+pub mod validation;
